feat(ui-kit): Sprint 4a Image split into RemoteImage, LocalImage, ImageRenderer
- Split monolithic Image.tsx into 4 files:
  - ImageRenderer.tsx — shared renderer logic
  - RemoteImage.tsx — URL-based images with lazyLoad
  - LocalImage.tsx — dynamic import with pendingLocal
  - Image.tsx — discriminated union entry point (RemoteImage | LocalImage)
- Update model/types.ts with ImageRendererOwnProps, ImageRendererProps
- Update index.ts to export RemoteImage, LocalImage, ImageRenderer
- Remove monolithic Image.tsx (was 450+ lines)

All validation passes: type-check, lint, 2254 tests, build, bundle 415.6 KiB
