+++ public/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + React + TS</title>
-    <script type="module" crossorigin src="/index.BzD1PLpt.js"></script>
+    <script type="module" crossorigin src="/index.Dr4Hqoae.js"></script>
     <link rel="modulepreload" crossorigin href="/rolldown-runtime.CbXtAM7H.js">
     <link rel="modulepreload" crossorigin href="/vendor.dcnawyM8.js">
     <link rel="modulepreload" crossorigin href="/i18n.2I-Qy6na.js">
